Reviews building process to allow build in one step. [#45]
This fixes all files that referenced Bridge.NET by its assembly .dll file
which generated cyclic dependency, resulting on a needed four step in order
to build the solution.

Now Bridge.NET is internally referred to as a 'Project Reference', so it
smartly tracks generated asm files before trying to build.

In the process, all projects affected by this change will call the 'deploy'
script in order to copy the generated files to the 'assemblies' directory.