Fixes a problem where files did not get '.js' extension on build. [#49]
If files had a '.' (dot) in the file name, Bridge's translator assumed
the files already had an extension and did not add the needed text '.js'.

This commit fixes that by defining a constant 'js' to the extension and then
performing string operations instead of path operation during the filename
extension evaluation.